add support for setting daterefs from sliders
This allows to set datarefs from sliders. If a 'dateref'
property is present, this is used instead of the command_{up,mid,down}
commands. The index parameter can be used to specify the index if
the dataref is of type array. scale can be used as to multiply the
value by 'scale' before writing it. The offset member can be used to
specify an offset that is added to the value.

Signed-off-by: Sven Schnelle <svens@stackframe.org>