Add akai mode to map_in_enc
The rotary encoder on the AKAI APC40 are a bit different then the
encoders currently supported by XMidiCtrl. Instead of returning the
Speed they where rotated with, the return an absolute Value between
0 - 127. The good thing is that they repeat the messages when at the
end of the range. That means we can emulate the up/down behaviour by:

0    -> encoder at minimum -> command_down
0x7f -> encoder at maximum -> command_up
previous value > current value: command_down
previous value < current value: command_up

This adds a mode parameter to the enc type, so a config entry looks
like this:

{ ch = 1, cc = 16,  type = "enc", mode="akai", command_up = "sim/autopilot/airspeed_up", command_down = "sim/autopilot/airspeed_down" },

Signed-off-by: Sven Schnelle <svens@stackframe.org>