Allow to configure value send in output messages
This adds a 'send_data' parameter that defines the value
sent in the midi out message. This is useful for note_on
commands that sets LEDs (like on Akai APC40), where the value
defines the color and whether the LED should blink or not.
If it is not set the default value of 127 is sent.

An example looks like this:

{ ch = 5, note_on = 53, send_data="1", type = "drf", dataref = "laminar/B738/annunciator/nose_gear_safe", value_on="1" }

Signed-off-by: Sven Schnelle <svens@stackframe.org>